Move Roots/RootFinder to poly package.
